sd3/tools/sd35m_cli_test/requirements.txt
# Inference stack
torch
torchvision
diffusers
transformers
accelerate
safetensors

# Image utils
Pillow
numpy
tqdm







